WeatherApp/requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
pytest==6.2.5
